requirements.txt
aiohttp==3.11.12
mcp==1.26.0
plexapi==4.18.0
python-dotenv==1.2.1
Requests==2.32.5
starlette==0.52.1
uvicorn==0.40.0
watchdog==6.0.0
